Fixed warnings
