+++ Firmware/Src/global_callbacks.cpp
@@ -32,6 +32,12 @@ volatile char uart_buf[UART_BUF_SIZE] = {0}; //
 std::vector<CommandHandler*> cmdHandlers;
 
 
+int _write(int file, char *ptr, int len)
+{
+  CDC_Transmit_FS(ptr, len);
+  return len;
+}
+
 std::vector<AdcHandler*> adcHandlers;
 void HAL_ADC_ConvCpltCallback(ADC_HandleTypeDef* hadc){
 	//Pulse braking mosfet if internal voltage is higher than supply. Conversion: V = 1/36
